Test: Refactor maximum_subarray tests
- Refactor maximum_subarray tests to be inside the same file. This repo
  will get crowded with multiple tests, so it may be a good idea to
  simplify the repo layout.
